cooldent/requirements.txt
Django==1.9.6
django-bootstrap-typeahead==1.1.5
django-braces==1.9.0
django-crispy-forms==1.6.0
django-highcharts==0.1.7
docutils==0.12
Pillow==4.1.0
simplejson==3.8.2
six==1.10.0
psycopg2==2.7.1
gunicorn==19.3.0